release: 0.1.0

+++ lib/yoomoney/models/line_item.rb
@@ -4,45 +4,37 @@ module Yoomoney
   module Models
     class LineItem < Yoomoney::Internal::Type::BaseModel
       # @!attribute description
-      #   Название товара или услуги (от 1 до 128 символов). Пользователь увидит его на
-      #   странице счета перед оплатой.
+      #   Описание товара или услуги.
       #
       #   @return [String]
       required :description, String
 
       # @!attribute price
-      #   Сумма в выбранной валюте.
+      #   Цена товара или услуги.
       #
       #   @return [Yoomoney::Models::MonetaryAmount]
       required :price, -> { Yoomoney::MonetaryAmount }
 
       # @!attribute quantity
-      #   Количество товара. Можно передать целое или дробное число. Разделитель дробной
-      #   части — точка, разделитель тысяч отсутствует, максимум три знака после точки.
-      #   Пример: 5.000
+      #   Количество.
       #
-      #   @return [Float]
-      required :quantity, Float
+      #   @return [Integer]
+      required :quantity, Integer
 
       # @!attribute discount_price
-      #   Сумма в выбранной валюте.
+      #   Цена со скидкой.
       #
       #   @return [Yoomoney::Models::MonetaryAmount, nil]
       optional :discount_price, -> { Yoomoney::MonetaryAmount }
 
       # @!method initialize(description:, price:, quantity:, discount_price: nil)
-      #   Some parameter documentations has been truncated, see
-      #   {Yoomoney::Models::LineItem} for more details.
+      #   @param description [String] Описание товара или услуги.
       #
-      #   Данные о товаре или услуге в корзине.
+      #   @param price [Yoomoney::Models::MonetaryAmount] Цена товара или услуги.
       #
-      #   @param description [String] Название товара или услуги (от 1 до 128 символов). Пользователь увидит его на ст
+      #   @param quantity [Integer] Количество.
       #
-      #   @param price [Yoomoney::Models::MonetaryAmount] Сумма в выбранной валюте.
-      #
-      #   @param quantity [Float] Количество товара. Можно передать целое или дробное число. Разделитель дробной ч
-      #
-      #   @param discount_price [Yoomoney::Models::MonetaryAmount] Сумма в выбранной валюте.
+      #   @param discount_price [Yoomoney::Models::MonetaryAmount] Цена со скидкой.
     end
   end
 end
